backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
supabase==2.9.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
httpx>=0.27.0
pydantic==2.6.3
lxml==5.1.0
playwright==1.42.0
